Plot display workaround. Explicitly set nodes red.
pip installed matplotlib does not come with tkagg backend for displaying the plot. I think an apt-get install might fix this, however, since all our other requirements are installed by pip, I am reluctant to make an exception for this. Instead, my proposed workaround is to simply save the plot and let the user choose how they wish to display the png file.

After looking at Victoria's webinar, I suspect Victoria's nodes default to red. That was not the case for my environment, hence the explicit declaration in the code.